// src/core/types.ts
export type Config = {
  minViews: number
  minDurationSeconds: number
  maxDurationSeconds: number
  maxQueueSize: number
  maxRequestsPerUser: number
  allowShorts: boolean
  allowLiveStreams: boolean
  contentMode: 'music' | 'any'
  regionCode: string
  fallbackPlaylist: FallbackPlaylist
}

export type FallbackPlaylist = {
  playlistId: string | null
  enabled: boolean
  shuffle: boolean
  repeat: boolean
}